Move stbt_core.pylint_plugin to _stbt
So pylint can import it without having to import the whole of
`stbt_core` (including cv2, etc.).
